requirements.txt
Flask
pymongo
python-dotenv
bcrypt
requests
